requirements.txt
kafka-python
django
dj-database-url
django-use-email-as-username
firebase_admin